requirements.txt
# Research MCP Server Requirements

# Core dependencies
mcp>=0.1.0
fastmcp>=0.9.0

# HTTP and API client
requests>=2.28.0
httpx>=0.24.0

# Data processing
python-dotenv>=1.0.0

# LLM support for enhanced parsing
openai>=1.0.0

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Optional: ArXiv search engine support
arxiv>=1.4.0

# Optional: Advanced NLP features for ArXiv engine
scikit-learn>=1.3.0

# Optional: For advanced NLP keyword extraction (if needed later)
# spacy>=3.5.0
# nltk>=3.8.0